Updated some fixes and updated theme of GUI
