+++ UVSim-Project/src/config.json
@@ -1,0 +1,4 @@
+{
+  "primary_color": "#4C721D",
+  "off_color": "#FFFFFF"
+}
